requirements.txt
Django==1.8.1
Django-Accounts==0.1
Fabric==1.10.1
antiorm==1.1.1
argparse==1.2.1
db==0.1.1
db-sqlite3==0.0.1
django-appconf==1.0.1
django-bootstrap-form==3.2
django-crispy-forms==1.4.0
django-debug-toolbar==1.3.0
django-staticfiles==1.2.1
django-user-accounts==1.0.2
ecdsa==0.13
paramiko==1.15.2
pinax-theme-bootstrap==5.7.3
pycrypto==2.6.1
pygraphviz==1.2
pytz==2015.2
raven==5.3.1
six==1.9.0
sqlparse==0.1.15
uWSGI==2.0.10
wsgiref==0.1.2
